==> 3.create_cellpainting_index.ipynb <==
import numpy as np
import pandas as pd

df_all = pd.read_csv("csv/processed_data_concat.csv")

df_all['Metadata_Plate'] = df_all['Metadata_Plate'].apply(str)
df_all['Metadata_Well'] = df_all['Metadata_Well'].str.upper()
df_all["SAMPLE_KEY"]= df_all[['Metadata_Plate', 'Metadata_Well']].agg('-'.join, axis=1)
df_all_small = pd.concat([df_all[['SAMPLE_KEY', 'Metadata_broad_sample']],df_all.iloc[:,17:-1]], axis=1)

df_smiles = pd.read_csv("csv/chemical_annotations.csv")
df_smiles = df_smiles.rename(columns={"BROAD_ID":"Metadata_broad_sample"})

from rdkit import Chem
from rdkit.Chem.inchi import MolToInchiKey

def smiles_to_mol(smiles):
    """ Convert a SMILES string to INCHIKEY
    """
    if not smiles:
        return ""
    try:
        m = Chem.MolFromSmiles(smiles)
    except:
        return ""
    inchikey = MolToInchiKey(m)
    return inchikey

df_merge = pd.merge(df_smiles[["Metadata_broad_sample", "CPD_SMILES"]], 
         df_all_small[df_all_small["Metadata_broad_sample"]!="DMSO"], 
         how = "right",
         on="Metadata_broad_sample")
df_merge.insert(0,'INCHIKEY','')
df_merge['INCHIKEY'] = df_merge['CPD_SMILES'].apply(smiles_to_mol)

df_merge = df_merge.drop(columns="Metadata_broad_sample")

df_merge

df_merge.to_csv("temp/cellpainting-index.csv", index=False)
