==> 5.process_duplicate_assay.ipynb <==
import pandas as pd

df = pd.read_csv('temp/chembl31-labels-assay-index.csv')
df = df.applymap(str)

counter = 1
while len(df[df.duplicated() == True])!=0:
    df.loc[df[df.duplicated() == True].index, :] = df.loc[df[df.duplicated() == True].index, :] + '_' + str(counter)
    counter += 1

df.to_csv("temp/chembl31-labels-assay-index-nondup.csv", index=False)
