==> 6.extract_assay_target.ipynb <==
import sqlite3
import pandas as pd

# Init path
chembl_db_path = "/home/son.ha/FSL_CP_DataPrep/sql/chembl_31_sqlite/chembl_31.db"

# Query
query_assay_target = "SELECT DISTINCT ay.assay_id, ay.chembl_id, t.chembl_id, t.target_type " \
                                "from assays ay, target_dictionary t " \
                                "where ay.tid = t.tid "

# Fetch 
db = sqlite3.connect(chembl_db_path)
cursor = db.cursor()
myresult = cursor.execute(query_assay_target).fetchall()


df_target = pd.DataFrame(myresult, columns=['ASSAY_ID', 'assay_chembl_id', 'target_chembl_id', 'target_type'])
df_target

df_cp_assay = pd.read_csv("temp/chembl31-labels-assay-index.csv")

df_cp_assay

df_merge = pd.merge(df_cp_assay, df_target, on="ASSAY_ID", how="left")
df_merge

df_assay_2 = pd.read_csv("temp/chembl31-labels-assay-index-nondup.csv")

df_merge["ASSAY_ID"] = df_assay_2["ASSAY_ID"]

df_merge['target_type'].value_counts()

df_merge.to_csv("output/assay_target_map.csv", index=False)
