==> 8.create_CP_feature_df.ipynb <==
## Make Cell Profiler Feature dataframe for only compounds which is also in ChEMBL

import pandas as pd
import numpy as np
import jsonlines

#Read df
df_compound = pd.read_csv('temp/cp_inchi_smiles_map.csv')
df_cp_profile = pd.read_csv('csv/processed_data_concat.csv')
df_chem = pd.read_csv('csv/chemical_annotations.csv')
df_chem = df_chem.rename(columns={"BROAD_ID":"Metadata_broad_sample"})

# Create a "SAMPLE_KEY" columns
def to_upper_add_dash(word):
    return("-"+word.upper())
df_cp_profile["SAMPLE_KEY"] = df_cp_profile["Metadata_Plate"].astype(str) + df_cp_profile["Metadata_Well"].apply(to_upper_add_dash)

df_cp_profile_small = pd.concat([df_cp_profile[["Metadata_Plate", "Metadata_Well", "SAMPLE_KEY", "Metadata_broad_sample"]], df_cp_profile.iloc[:,17:-1]], axis=1)

df_merge_all = pd.merge(df_chem[['CPD_SMILES','Metadata_broad_sample']], df_cp_profile_small, on="Metadata_broad_sample", how="right")

df_merge_cp_only = pd.merge(df_compound, df_merge_all, on="CPD_SMILES", how="left")

df_merge_cp_only.sort_values(by=["Metadata_Plate", "Metadata_Well"]).to_csv("output/CP_feature_df.csv", index=False)

df_merge_cp_only.sort_values(by=["Metadata_Plate", "Metadata_Well"])[["INCHIKEY", "CPD_SMILES", "Metadata_broad_sample", "Metadata_Plate", "Metadata_Well", "SAMPLE_KEY"]].to_csv("temp/cp_inchi_smiles_sample_key_map.csv", index=False)
