==> 12.create_final_label_df.ipynb <==
import pandas as pd
import numpy as np
from tqdm import tqdm
from scipy.io import mmread
import os

# Each duplicate well is 1 datapoint

image_folder = "/mnt/scratch/Son_cellpainting/my_cp_images/"
jsonl_folder = "/home/son.ha/FSL_CP_DataPrep/jsonl/"

### Load Data

# Features
CP_df = pd.read_csv('output/norm_CP_feature_df.csv')
FP_df = pd.read_csv('output/norm_ECFP_feature_df.csv')
RDKit_df = pd.read_csv('output/norm_RDKit_feature_df.csv')

# Assert that inchikeys should be the same and have the same order
assert CP_df['INCHIKEY'].equals(RDKit_df['INCHIKEY'])
assert RDKit_df['INCHIKEY'].equals(FP_df['INCHIKEY'])

# Assay Labels
assay = pd.read_csv('output/label_df.csv', index_col=0)
labels_with_dupl = pd.merge(FP_df[['INCHIKEY', 'CPD_SMILES', 'SAMPLE_KEY']].set_index('INCHIKEY'), assay, left_index=True, right_index=True)

### Find out which views are missing from each well

view_counts = {}
for file in os.listdir(image_folder):
    if file.endswith('.npz'):
        sample_key = file[0:9]
        view = file[-5:-4]
        if sample_key not in view_counts.keys():
            view_counts[sample_key] = [view]
        else:
            view_counts[sample_key].append(view)
            view_counts[sample_key].sort()

count_missing = 0
for sample_key in view_counts.keys():
    if len(view_counts[sample_key]) != 6:
        #print(sample_key, view_counts[sample_key])
        missing_num = 6-len(view_counts[sample_key])
        count_missing += missing_num
print(f"There are {count_missing} views missing across all wells") 

### Add a views column to the label dataframe

def _retrieve_view(sample_key):
    """Helper function. Retrive a list of available views given sample key"""
    return('_'.join(view_counts[sample_key]))

labels_with_dupl['VIEWS'] = labels_with_dupl['SAMPLE_KEY'].apply(_retrieve_view)

labels_with_dupl.loc[labels_with_dupl.loc[:, ['CPD_SMILES', 'SAMPLE_KEY', labels_with_dupl.columns[3]]].replace(-1, np.NaN).dropna().index[1]]

labels_with_dupl.columns[2:-1]

def _sample_key_to_num_row_cp(sample_key):
    '''Helper function for pandas apply
    Input a sample_key, return the index of the feature dataframe of that sample_key'''
    return(int(CP_df.loc[CP_df['SAMPLE_KEY']==sample_key].index[0]))

df_for_each_assay = []
for column in tqdm(labels_with_dupl.columns[2:-1]):
    # Iterate through each column, remove NA
    temp_df = labels_with_dupl.loc[:, ['CPD_SMILES', 'SAMPLE_KEY', 'VIEWS', column]].replace(-1, np.NaN).dropna().copy(deep=True)
    temp_df['NUM_ROW_CP_FEATURES'] = temp_df['SAMPLE_KEY'].apply(_sample_key_to_num_row_cp)
    temp_df['ASSAY'] = str(column)
    temp_df = temp_df.rename(columns={column:'LABEL'})
    df_for_each_assay.append(temp_df)

final_label_df = pd.concat(df_for_each_assay)
final_label_df = final_label_df.reset_index()
final_label_df = final_label_df.rename(columns={'index':'INCHIKEY'})
final_label_df['ASSAY'] = final_label_df['ASSAY'].astype(str)
final_label_df

final_label_df.to_csv('output/FINAL_LABEL_DF.csv', index=False)
